Multi-Language-Translation/requirements.txt
googletrans==4.0.0rc1
streamlit==1.31.1
python-docx=1.1.0